Move evaluateGuard to guards.ts and simplify
